examples/python/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-azure-app-config==0.0.1-alpha.1699945013+97b0e04c
